# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  commands.CreateServiceOrderRequest:
    properties:
      notes:
        type: string
      requester_document:
        type: string
      vehicle_plate:
        type: string
    required:
    - requester_document
    - vehicle_plate
    type: object
...
  commands.CreateVehicleRequest:
    properties:
      brand:
        type: string
      model:
        type: string
      plate:
        type: string
      requester_document:
        type: string
      year:
        type: integer
    required:
    - brand
    - model
    - plate
    - requester_document
    - year
    type: object
...
  commands.MessageResponse:
    properties:
      message:
        type: string
    type: object
  commands.PartItemRequest:
    properties:
      manufacturer_code:
        type: string
      quantity:
        type: integer
    required:
    - manufacturer_code
    - quantity
    type: object
  commands.PartItemResponse:
    properties:
      description:
        type: string
      part_id:
        type: string
      quantity:
        type: integer
      unit_price:
        type: number
    type: object
...
  commands.RegisterRequest:
    properties:
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  commands.ServiceItemRequest:
    properties:
      code:
        type: integer
    required:
    - code
    type: object
  commands.ServiceItemResponse:
    properties:
      description:
        type: string
      price:
        type: number
      service_id:
        type: string
    type: object
  commands.ServiceOrderResponse:
    properties:
      code:
        type: integer
      created_at:
        type: string
      finished_at:
        type: string
      id:
        type: string
      notes:
        type: string
      parts:
        items:
          $ref: '#/definitions/commands.PartItemResponse'
        type: array
      requester_id:
        type: string
      services:
        items:
          $ref: '#/definitions/commands.ServiceItemResponse'
        type: array
      started_at:
        type: string
      status:
        $ref: '#/definitions/entities.OrderStatus'
      total_amount:
        type: number
      updated_at:
        type: string
      vehicle_id:
        type: string
    type: object
  commands.ServiceResponse:
    properties:
      code:
        type: integer
      created_at:
        type: string
      description:
        type: string
      duration_min:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
...
  commands.UpdateServiceOrderRequest:
    properties:
      notes:
        type: string
    type: object
  commands.UpdateServiceRequest:
    properties:
      code:
        type: integer
      description:
        type: string
      duration_min:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
...
  commands.UpdateStatusRequest:
    properties:
      parts:
        items:
          $ref: '#/definitions/commands.PartItemRequest'
        type: array
      services:
        items:
          $ref: '#/definitions/commands.ServiceItemRequest'
        type: array
      status:
        $ref: '#/definitions/entities.OrderStatus'
    required:
    - status
    type: object
  commands.UpdateVehicleRequest:
    properties:
      brand:
        type: string
      model:
        type: string
      plate:
        type: string
      year:
        type: integer
    type: object
  commands.VehicleResponse:
    properties:
      brand:
        type: string
      created_at:
        type: string
      id:
        type: string
      model:
        type: string
      plate:
        type: string
      requester_id:
        type: string
      updated_at:
        type: string
      year:
        type: integer
    type: object
  entities.OrderStatus:
    enum:
    - received
    - in_diagnosis
    - awaiting_approval
    - in_execution
    - finished
    - delivered
    type: string
    x-enum-varnames:
    - StatusReceived
    - StatusInDiagnosis
    - StatusAwaitingApproval
    - StatusInExecution
    - StatusFinished
    - StatusDelivered
...
host: localhost:8080
info:
  contact: {}
  description: Sistema Integrado de Atendimento e Execucao de Servicos — Oficina Mecanica
  title: Workshop API
  version: "1.0"
paths:
...
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do usuario
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commands.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/commands.MessageResponse'
      summary: Registrar usuario
      tags:
      - auth
...
  /service-orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/commands.ServiceOrderResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Listar ordens de servico
      tags:
      - service-orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados da OS
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commands.CreateServiceOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/commands.ServiceOrderResponse'
      security:
      - BearerAuth: []
      summary: Criar ordem de servico
      tags:
      - service-orders
  /service-orders/{id}:
    delete:
      parameters:
      - description: ServiceOrder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Deletar ordem de servico
      tags:
      - service-orders
    get:
      parameters:
      - description: ServiceOrder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commands.ServiceOrderResponse'
      security:
      - BearerAuth: []
      summary: Buscar ordem de servico por ID
      tags:
      - service-orders
    put:
      consumes:
      - application/json
      parameters:
      - description: ServiceOrder ID
        in: path
        name: id
        required: true
        type: string
      - description: Dados para atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commands.UpdateServiceOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commands.ServiceOrderResponse'
      security:
      - BearerAuth: []
      summary: Atualizar ordem de servico
      tags:
      - service-orders
  /service-orders/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: ServiceOrder ID
        in: path
        name: id
        required: true
        type: string
      - description: Novo status (services e parts obrigatorios para awaiting_approval)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commands.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Atualizar status da ordem de servico
      tags:
      - service-orders
  /service-orders/code/{code}:
    get:
      parameters:
      - description: Codigo da OS
        in: path
        name: code
        required: true
        type: integer
      - description: CPF ou CNPJ do cliente
        in: query
        name: document
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commands.ServiceOrderResponse'
      summary: Consultar OS pelo codigo (endpoint publico do cliente)
      tags:
      - service-orders-requester
...
  /services/{id}:
    delete:
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Deletar servico
      tags:
      - services
    get:
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commands.ServiceResponse'
      security:
      - BearerAuth: []
      summary: Buscar servico por ID
      tags:
      - services
    put:
      consumes:
      - application/json
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Dados para atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commands.UpdateServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commands.ServiceResponse'
      security:
      - BearerAuth: []
      summary: Atualizar servico
      tags:
      - services
  /vehicles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/commands.VehicleResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Listar veiculos
      tags:
      - vehicles
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do veiculo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commands.CreateVehicleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/commands.VehicleResponse'
      security:
      - BearerAuth: []
      summary: Cadastrar veiculo
      tags:
      - vehicles
  /vehicles/{id}:
    delete:
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Deletar veiculo
      tags:
      - vehicles
    get:
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Buscar veiculo por ID
      tags:
      - vehicles
    put:
      consumes:
      - application/json
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: string
      - description: Dados para atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commands.UpdateVehicleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commands.VehicleResponse'
      security:
      - BearerAuth: []
      summary: Atualizar veiculo
      tags:
      - vehicles
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"